requirements.txt

streamlit
python-dotenv
pymupdf
langchain
sentence_transformers
chromadb
bertopic
hdbscan
scikit_learn
keybert
torch
transformers